test: Implement full DB module integration tests
These tests verify that all exported DB functions actually work on
a real database and have expected properties.

- Querying a non-existent character should not throw or return an error,
  just a Nothing. This also allows us to check that connecting to the DB
  at all works.
- Creating a character should work.
- Retrieving the character we created should work.
- Creating a duplicate character should report `Left (DBError UniqueConstraintViolated)`
- Deleting a character should work (this also cleans up the test
  session; this is currently the only use for the deleteCharacter function)
- We double check deleting the character works by querying it again and
  expecting a Nothing.

Currently the tests are run against the production database, because
it's safe to do so. If eventually a test database is required, extra
care will be required to ensure it has the same properties as the
production database. At that time, a migration manager will be
necessary. Unfortunately, Opaleye can't help us here.

A clever shortcircuiting mechanism is included to abort all operations
if one fails.