feat: Character/Random Character page V1 done
The Character/Random Character page is pretty much done.

It still uses dummy data for now, for testability with `elm reactor`.

Initially, the plan was to style description blocks differently
depending on what theme they cover, but I can't find any sensible design
for that. In line with the minimialistic feel of the whole site, let's
just style all of them the same.

This means the backend sends a lot of useless information:
- Description block breakdown
- Abilities breakdown (this is also not used)

However, let's keep sending this, as it is still interesting information
that either might be used in a future version or by other hypothetical
consumers of the API.

Other notable changes:

FE:
- page updates now take Ctx
- Ctx can be used to change the URL from any page
- The route that the user was previously on can be queried from the Ctx,
  allowing to change certain texts depending on where the user comes from.
- Using "Lato" web font
- Palette module has been made DRYer

BE:
- Improved description markdown formatting