fix: decode name when picking it up from URL
The "name" input was previously placed in the URL
then taken again from there, and it seems the Elm
parser was not "percent decoding" it despite the documentation implying
it should. Instead, we do that manually.

This was causing e.g. spaces and accents to appear percent-encoded e.g.
"%20" instead of " ".